8248458: CSupport.toJavaStringInternal should be moved SharedUtils
